refactor(web): finish core/ reorganisation from structure-refactor-plan
- Move OnboardingWizard.tsx → core/onboarding/ (update relative imports)
- Move ProfilePage.test.tsx → core/profile/
- Remove ProfilePage.tsx re-export barrel (App.tsx now imports directly)
- Update eslint.config.js no-raw-local-storage ignores for new paths
