--- apps/web/src/core/ProfilePage.tsx
@@ -1,1 +1,0 @@
-export { ProfilePage } from "./profile/ProfilePage";
