+++ eslint.config.js
@@ -209,8 +209,7 @@ export default [
       "apps/web/src/core/hub/HubDashboard.tsx",
       "apps/web/src/core/hub/HubReports.tsx",
       "apps/web/src/core/hub/HubSearch.tsx",
-      "apps/web/src/core/OnboardingWizard.tsx",
-      "apps/web/src/core/ProfilePage.tsx",
+      "apps/web/src/core/onboarding/OnboardingWizard.tsx",
       "apps/web/src/core/insights/TodayFocusCard.tsx",
       "apps/web/src/core/observability/analytics.ts",
       "apps/web/src/core/app/pwaAction.ts",
